Refactor ExchangeResumeScreen hooks and handle errors
